MAINT: remove full_output kwarg from helper method BUG: fix behaviour when cube and geometry don't overlap TST: move tests from #1134 to new unit test structure
